+++ src/components/componentEmoji/CHANGELOG.md
@@ -1,5 +1,11 @@
 # component-emoji
 
+## 0.0.2
+
+### Patch Changes
+
+- Updating emojis to return a black and white version of the emoji for a clean UI.
+
 ## 0.0.1
 
 ### Patch Changes